bugfix: error handling when sending audio
Two bugs, both related to how errors are handled when sending audio to the
server have been fixed.

One bug was a typo, where `nil` was returned instead of `err`.

The other bug was related to stream.Send failing with an EOF.  The EOF only
informs the client that the stream has ended; the actual error status is
retrieved from the Recv call.  Hence we don't send the io.EOF on the error
channel and let the actual status obtained from Recv be returned as the main error.