[wip] api improvements (breaking changes)
This commit introduces the following changes to the protobuf API:

1. Remove dependency on protobuf.Empty: An empty message is defined in
this proto file instead.

2. Remove dependency on protobuf.Duration: Timestamps are now reported
in milliseconds.  Using the protobuf.Duration message types was not very
convenient in some client languages.

3. Non-streaming Recognize() rpc call removed:  This was almost never
being used, and having it in was causing confusion to new users of the
API.

4. A new AudioFormat field is introduced in RecognitionConfig messages:
previously, the AudioEncoding enum was mixing different aspects of audio
specification (codec, samplerate, raw/headered formats).  The new config
would require users to provide either a "Headered" type, so that the
server can deduce audio parameters from their headers, or provide all
details required to process raw audio (sample rate, bit-depth,
endianness, encoding and number of channels).  It has also been
clarified in the documentation that OPUS format is supported with ogg
headers only.

5. The idle_timeout field is removed from RecognitionConfig because it
has been found to be confusing to use without any particular benefit to
clients.  Servers can/will still set idle timeouts as necessary.

6. A new field initial_time_offset has been introduced to
RecognitionConfig.  If a recognition session was interrupted, and a
client would like to restart recognizing audio starting at a certain
offset with a new recognizer, it would be desired that the timestamps
published in the results conform to the original audio stream.

7. The audio_channels field in RecognitionConfig has been renamed to
selected_audio_channels, and its behavior is changed: previously, when
this field was unset, only the first channel of the input audio was
decoded.  The new behavior is that if the field is unset, all input
channels would be decoded.  However, clients can choose to decode only
some of the channels by setting those indices in this field.